Deploying with Terminal updated
The current text on gridsome.com and zeit.co is not describing the all details about deploying to Now. 

Following it, you have to build it locally, then go to `dist` folder, type `now`. Thus it won't find the `now.json` file so you are getting a project named `dist` in Now.

In the new text, it is described how to configure the build process so that you could just type `now` and get the project with the proper name in Now.