Update guide-search.md
A quick answer for search. It's kinda meh, but quick. I can expand this with screen shots and a sample project too, but I'd also like to talk about a client-side local option like Lunr, and you may have seen my issues related to that. If you don't want a quick fix that's fine too, it took me all of 5 minutes to write. :)