Show searchbar in nav when in docs on mobile
Lets you search docs on much smaller screens from the header. But only if the current path starts with "/docs". With this it's easier to type and see the immediate results at once in portrait mode - very hard in landscape because the onscreen keyboard and header takes up all the visual area.

Elsewhere this reverts to normal behaviour to show the external social links icons if screen is small and hide the search bar. 

Agolia search automatically adjusts the ltr/rtl direction based on screen size and current input position at render and it seems to fit fine on the few screen devices I tried with the slight alteration to tie search results min-width to screen size.