Update client-side-data.md (#170)
* Update client-side-data.md

Update the fetching from a local YAML file explanation to match the voice on the rest of the docs. Updated the example using object destructuring.

Added fetching from a local JSON file. I know it mirrors the YAML file example, but I wanted to call it out explicitly as is done in other places in the docs.

* Update client-side-data.md

fix import in example for JSON