Add directions for markdown files to fetching-data.md
I've struggled a couple of times now with adding Markdown file support to my Gridsome site, and always had to rely on third-party blog posts. I realized the info _is_ all here in the docs, but in a few different places, so I thought adding those directions here all in one place to the official documentation would be a great way to contribute to this excellent project. (And it would be my first open-source contribution!)

More than open to any and all feedback and revisions. Thanks for maintaining such a great project!