Adding a CircleCI build to lint markdown (#466)
* Adding a CircleCI build to lint markdown

We don't have a very consistent style for Markdown currently. With many contributors writing according to their preferences, it can be confusing which Markdown conventions to follow. Adding a linter will help us write in a consistent style.

This PR starts with a very loose ruleset for markdownlint to minimize noise. We're only fixing a single rule (hard tabs). By tackling each rule at a time, we can keep the PRs small and tightly-scoped.

* Switching from CircleCI to Github Actions